Improved the README.md
Added an Idea section to describe the goal of the project.
Added Current Status section to describe the current state of the
project and it's modules
Added Current Status / Supported Statements section to describe the
supported import types, their status, and how they work; also lists
caviots and important points / how to use.
Added ToDo section to list things that need to be done in order of
importance / requirement.